+++ tests/inputs-editorconfig/space-2.lua
@@ -1,0 +1,5 @@
+local foo = {
+	a = 1,
+}
+
+local bar = ""