+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.17.0] - 2023-03-11
+
 ### Added
 
 - Added support for "sort requires", which sorts top-level statements of the form `local NAME = require(EXPR)` lexicographically on `NAME`.
@@ -643,7 +645,8 @@ This feature is enabled by default, it can be disabled using `--no-editorconfig`
 
 Initial alpha release
 
-[unreleased]: https://github.com/JohnnyMorganz/StyLua/compare/v0.16.1...HEAD
+[unreleased]: https://github.com/JohnnyMorganz/StyLua/compare/v0.17.0...HEAD
+[0.17.0]: https://github.com/JohnnyMorganz/StyLua/releases/tag/v0.17.0
 [0.16.1]: https://github.com/JohnnyMorganz/StyLua/releases/tag/v0.16.1
 [0.16.0]: https://github.com/JohnnyMorganz/StyLua/releases/tag/v0.16.0
 [0.15.3]: https://github.com/JohnnyMorganz/StyLua/releases/tag/v0.15.3